Remove `change_plan` adapter
