+++ lib/plausible/teams/adapter/read/billing.ex
@@ -5,13 +5,6 @@ defmodule Plausible.Teams.Adapter.Read.Billing do
   use Plausible.Teams.Adapter
   use Plausible
 
-  def change_plan(user, new_plan_id) do
-    switch(user,
-      team_fn: &Plausible.Teams.Billing.change_plan(&1, new_plan_id),
-      user_fn: &Plausible.Billing.change_plan(&1, new_plan_id)
-    )
-  end
-
   def latest_enterprise_plan_with_prices(user, customer_ip) do
     switch(user,
       team_fn: &Plausible.Teams.Billing.latest_enterprise_plan_with_price(&1, customer_ip),
